feat: combine useRouterTrade & useRoutingAPITrade
